Added map name support on events page
Added support for map names in the events controller. I've also moved
the compairing functionality into the mapsFactory.
I've also created a new compairing function which uses a shortened
version of the ID provided. This is used because the ID provided by
ANET seems to be randomly generated and unique. Because of this, using
a smaller subset of the ID provided provides quicker compairing.